+++ src/Make_morph.mak
@@ -74,6 +74,7 @@ SRC =	arabic.c						\
 	message.c						\
 	misc1.c							\
 	misc2.c							\
+	mouse.c							\
 	move.c							\
 	normal.c						\
 	ops.c							\
