+++ src/proto.h
@@ -175,6 +175,7 @@ char_u *vim_strpbrk(char_u *s, char_u *charset);
 /* Use our own qsort(), don't define the prototype when not used. */
 void qsort(void *base, size_t elm_count, size_t elm_size, int (*cmp)(const void *, const void *));
 #endif
+# include "mouse.pro"
 # include "move.pro"
 # include "mbyte.pro"
 # include "normal.pro"
