+++ src/proto/term.pro
@@ -2,8 +2,6 @@
 guicolor_T termgui_get_color(char_u *name);
 guicolor_T termgui_mch_get_rgb(guicolor_T color);
 int set_termname(char_u *term);
-void set_mouse_termcode(int n, char_u *s);
-void del_mouse_termcode(int n);
 void getlinecol(long *cp, long *rp);
 int add_termcap_entry(char_u *name, int force);
 int term_is_8bit(char_u *name);
@@ -47,9 +45,6 @@ void may_req_termresponse(void);
 void may_req_ambiguous_char_width(void);
 void may_req_bg_color(void);
 int swapping_screen(void);
-void setmouse(void);
-int mouse_has(int c);
-int mouse_model_popup(void);
 void scroll_start(void);
 void cursor_on_force(void);
 void cursor_on(void);
@@ -64,6 +59,7 @@ void clear_termcodes(void);
 void add_termcode(char_u *name, char_u *string, int flags);
 char_u *find_termcode(char_u *name);
 char_u *get_termcode(int i);
+void del_termcode(char_u *name);
 void set_mouse_topline(win_T *wp);
 int check_termcode(int max_offset, char_u *buf, int bufsize, int *buflen);
 void term_get_fg_color(char_u *r, char_u *g, char_u *b);
